ai_service/requirements.txt
flask
requests
chromadb
sentence-transformers
vaderSentiment
python-dotenv
huggingface-hub
supabase
gunicorn
